ml/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.10
scikit-learn==1.7.1
scikit-surprise==1.1.4
mlflow==3.1.1
numpy==1.26.4